+++ CHANGELOG.md
@@ -30,6 +30,9 @@ The following changes have been completed, and will be included in the next tagg
   * DataPath
     - added missing `MakeDataPathOptions`
     - added missing `MAKE_DATA_PATH_DEFAULT_OPTIONS`
+    - `makeDataPath()` options are now compatible with the `TypeGuaranteeOptions`
+  * HttpStatusCode
+    - `makeHttpStatusCode()` options are now compatible with the `TypeGuaranteeOptions`
 
 ## v0.1.8